Extract VCS to class
vcs: extract interface bo abstract base class;
vcs: add implementation for git repository;
pcs: add helper for getting the vcs object;
various: fix minor issues through the codebase.
