+++ perun/check/factory.py
@@ -14,7 +14,6 @@ from typing import Any, Iterable, Protocol, TYPE_CHECKING
 # Third-Party Imports
 
 # Perun Imports
-from perun import vcs
 from perun.logic import config, pcs, runner, store
 from perun.select.abstract_base_selection import AbstractBaseSelection
 from perun.check.methods import (
@@ -145,7 +144,7 @@ def degradation_in_minor(
     :returns: list of found changes
     """
     selection: AbstractBaseSelection = pcs.selection()
-    minor_version_info = vcs.get_minor_version_info(minor_version)
+    minor_version_info = pcs.vcs().get_minor_version_info(minor_version)
 
     # Precollect profiles for all near versions
     pre_collect_profiles(minor_version_info)
@@ -188,7 +187,7 @@ def degradation_in_history(head: str) -> list[tuple[DegradationInfo, str, str]]:
     detected_changes = []
     version_selection: AbstractBaseSelection = pcs.selection()
     with log.History(head) as history:
-        for minor_version in vcs.walk_minor_versions(head):
+        for minor_version in pcs.vcs().walk_minor_versions(head):
             history.progress_to_next_minor_version(minor_version)
             newly_detected_changes = []
             if version_selection.should_check_version(minor_version):
