+++ perun/select/whole_repository_selection.py
@@ -12,7 +12,7 @@ from typing import Iterator
 # Third-Party Imports
 
 # Perun Imports
-from perun import vcs
+from perun.logic import pcs
 from perun.profile import helpers as profile_helpers
 from perun.profile.factory import Profile
 from perun.profile.helpers import ProfileInfo
@@ -55,7 +55,7 @@ class WholeRepositorySelection(AbstractBaseSelection):
         :param target_version: analysed target version
         :return: both parents of the version
         """
-        return [vcs.get_minor_version_info(parent) for parent in target_version.parents]
+        return [pcs.vcs().get_minor_version_info(parent) for parent in target_version.parents]
 
     def get_profiles(
         self, target_version: MinorVersion, target_profile: Profile
@@ -69,7 +69,7 @@ class WholeRepositorySelection(AbstractBaseSelection):
         """
         analysis_worklist = []
         for parent in target_version.parents:
-            parent_info = vcs.get_minor_version_info(parent)
+            parent_info = pcs.vcs().get_minor_version_info(parent)
             for profile_info in profile_helpers.load_list_for_minor_version(parent):
                 if profile_info.is_compatible_with_profile(target_profile):
                     analysis_worklist.append((parent_info, profile_info))
@@ -87,7 +87,7 @@ class WholeRepositorySelection(AbstractBaseSelection):
         :return: iterator of pairs of minor versions and their corresponding profiles;
             each profile is compatible with target profile
         """
-        for minor_version in vcs.walk_minor_versions(target_version.checksum):
+        for minor_version in pcs.vcs().walk_minor_versions(target_version.checksum):
             yield from self.get_profiles(minor_version, target_profile)
 
     def get_skeleton(self, target_version: MinorVersion) -> Iterator[MinorVersion]:
@@ -96,7 +96,7 @@ class WholeRepositorySelection(AbstractBaseSelection):
         :param target_version: analysed target version (the skeleton head)
         :return: iterator of minor versions
         """
-        yield from vcs.walk_minor_versions(target_version.checksum)
+        yield from pcs.vcs().walk_minor_versions(target_version.checksum)
 
     def find_nearest(self, target_version: MinorVersion) -> MinorVersion:
         """Returns the identity: the target version itself
