+++ perun/vcs/meson.build
@@ -2,7 +2,9 @@ perun_vcs_dir = perun_dir / 'vcs'
 
 perun_vcs_files = files(
     '__init__.py',
-    'git.py',
+    'abstract_repository.py',
+    'git_repository.py',
+    'vcs_kit.py'
 )
 
 py3.install_sources(
