+++ tests/conftest.py
@@ -21,7 +21,6 @@ import perun.logic.commands as commands
 import perun.utils.decorators as decorators
 import perun.utils.streams as streams
 import perun.utils.metrics as metrics
-import perun.vcs as vcs
 
 import perun.testing.utils as test_utils
 
@@ -261,7 +260,7 @@ def pcs_with_degradations():
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     # Populate repo with commits
     repo = git.Repo(pcs_path)
@@ -317,7 +316,7 @@ def pcs_single_prof(stored_profile_pool):
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     # Populate repo with commits
     repo = git.Repo(pcs_path)
@@ -365,7 +364,7 @@ def pcs_full(stored_profile_pool):
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     # Populate repo with commits
     repo = git.Repo(pcs_path)
@@ -418,7 +417,7 @@ def pcs_full_no_prof():
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     # Populate repo with commits
     repo = git.Repo(pcs_path)
@@ -458,7 +457,7 @@ def pcs_with_empty_git():
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     yield pcs
 
@@ -477,7 +476,7 @@ def pcs_with_root():
     commands.init_perun_at(pcs_path, False, {"vcs": {"url": "../", "type": "git"}})
 
     # Initialize git
-    vcs.init({})
+    pcs.vcs().init({})
 
     # Populate repo with commits
     repo = git.Repo(pcs_path)
