+++ tests/test_select.py
@@ -7,7 +7,7 @@ import pytest
 # Third-Party Imports
 
 # Perun Imports
-from perun import vcs
+from perun.logic import pcs
 from perun.profile import helpers as profile_helpers
 from perun.select.abstract_base_selection import AbstractBaseSelection
 from perun.select.whole_repository_selection import WholeRepositorySelection
@@ -22,10 +22,10 @@ def test_base_select():
 def test_whole_repository(pcs_with_degradations):
     git_repo = git.Repo(pcs_with_degradations.get_vcs_path())
 
-    head = vcs.get_minor_version_info(str(git_repo.head.commit))
-    parent = vcs.get_minor_version_info(head.parents[0])
-    second_parent = vcs.get_minor_version_info(head.parents[1])
-    root = vcs.get_minor_version_info(parent.parents[0])
+    head = pcs.vcs().get_minor_version_info(str(git_repo.head.commit))
+    parent = pcs.vcs().get_minor_version_info(head.parents[0])
+    second_parent = pcs.vcs().get_minor_version_info(head.parents[1])
+    root = pcs.vcs().get_minor_version_info(parent.parents[0])
 
     head_profile = profile_helpers.load_list_for_minor_version(head.checksum)[0].load()
     root_profile = profile_helpers.load_list_for_minor_version(root.checksum)[0].load()
