+++ tests/test_vcs.py
@@ -1,11 +1,17 @@
 """Basic tests for checking the correctness of the VCS modules"""
 
+# Standard Imports
 import git
-import os
 import operator
+import os
+import pytest
 
-import perun.vcs as vcs
-import perun.logic.store as store
+# Third-Party Imports
+
+# Perun Imports
+from perun.vcs import vcs_kit
+from perun.vcs.abstract_repository import AbstractRepository
+from perun.logic import pcs, store
 
 
 def test_on_empty_git(pcs_with_empty_git):
@@ -14,7 +20,7 @@ def test_on_empty_git(pcs_with_empty_git):
     Excepts handled errors and not uncaught shit
     """
     # Assert that when we walk with bogus head then nothing is returned
-    assert len(list(vcs.walk_minor_versions(""))) == 0
+    assert len(list(pcs.vcs().walk_minor_versions(""))) == 0
 
 
 def test_major_versions(pcs_full_no_prof):
@@ -25,22 +31,22 @@ def test_major_versions(pcs_full_no_prof):
     git_config_parser = git.config.GitConfigParser()
     git_default_branch_name = git_config_parser.get_value("init", "defaultBranch", "master")
 
-    major_versions = list(vcs.walk_major_versions())
+    major_versions = list(pcs.vcs().walk_major_versions())
 
     assert len(major_versions) == 1
     major_version = major_versions[0]
     assert major_version.name == git_default_branch_name
     assert store.is_sha1(major_version.head)
 
-    head_major = vcs.get_head_major_version()
+    head_major = pcs.vcs().get_head_major_version()
     assert not store.is_sha1(str(head_major))
     assert str(head_major) == git_default_branch_name
 
-    prev_commit = vcs.get_minor_version_info(vcs.get_minor_head()).parents[0]
+    prev_commit = pcs.vcs().get_minor_version_info(pcs.vcs().get_minor_head()).parents[0]
     git_repo = git.Repo(pcs_full_no_prof.get_vcs_path())
     git_repo.git.checkout(prev_commit)
     # Try to detach head
-    head_major = vcs.get_head_major_version()
+    head_major = pcs.vcs().get_head_major_version()
     assert store.is_sha1(head_major)
 
 
@@ -51,43 +57,43 @@ def test_saved_states(pcs_full_no_prof):
     """
 
     # Is not dirty
-    assert not vcs.is_dirty()
+    assert not pcs.vcs().is_dirty()
 
     with open("file2", "r+") as write_handle:
         previous_state = write_handle.readlines()
         write_handle.write("hello")
 
     # Should be dirty
-    assert vcs.is_dirty()
+    assert pcs.vcs().is_dirty()
 
     # The changes should be cleared
-    with vcs.CleanState():
-        assert not vcs.is_dirty()
+    with vcs_kit.CleanState():
+        assert not pcs.vcs().is_dirty()
 
         with open("file2", "r") as read_handle:
             new_state = read_handle.readlines()
         assert new_state == previous_state
 
-    head = vcs.get_minor_head()
-    minor_versions = list(map(operator.attrgetter("checksum"), vcs.walk_minor_versions(head)))
+    head = pcs.vcs().get_minor_head()
+    minor_versions = list(map(operator.attrgetter("checksum"), pcs.vcs().walk_minor_versions(head)))
 
     with open("file2", "w") as write_handle:
         write_handle.write("".join(previous_state))
 
-    with vcs.CleanState():
+    with vcs_kit.CleanState():
         # Now try checkout for all of the stuff
-        vcs.checkout(minor_versions[1])
+        pcs.vcs().checkout(minor_versions[1])
         tracked_files = os.listdir(os.getcwd())
         assert set(tracked_files) == {".perun", ".git", "file1", "file2"}
 
     # Test that the head was not changed and kept unchanged by CleanState
-    assert vcs.get_minor_head() == head
+    assert pcs.vcs().get_minor_head() == head
     # Assert that save state is not used if the dir is not dirty:w
-    assert not vcs.is_dirty() and not vcs.save_state()[0]
+    assert not pcs.vcs().is_dirty() and not pcs.vcs().save_state()[0]
 
     # Test saving detached head state
-    vcs.checkout(minor_versions[1])
-    saved, _ = vcs.save_state()
+    pcs.vcs().checkout(minor_versions[1])
+    saved, _ = pcs.vcs().save_state()
     assert not saved
 
 
@@ -96,5 +102,10 @@ def test_diffs(pcs_full_no_prof):
 
     Expecting correct behaviour and no error
     """
-    diff = vcs.minor_versions_diff("HEAD", "HEAD~1")
+    diff = pcs.vcs().minor_versions_diff("HEAD", "HEAD~1")
     assert diff != ""
+
+
+def test_abstract_base():
+    with pytest.raises(TypeError):
+        _ = AbstractRepository()
